refactor: 뽀모도로 중복 함수, 타입 리팩토링
