+++ src/features/pomodoro/model/types.ts
@@ -1,0 +1,8 @@
+export interface PomodoroSyncData {
+  type: 'pomodoro-update';
+  phase: 'study' | 'break';
+  remainingTime: number;
+  isRunning: boolean;
+  totalCycles: number;
+  timestamp: number;
+}
