+++ src/features/pomodoro/ui/PomodoroController.tsx
@@ -4,6 +4,7 @@ import {
   STUDY_TIME,
   BREAK_TIME,
 } from '../model/usePomodoroStore';
+import { formatTime } from '../model/utils';
 import { CircularProgress } from './CircularProgress';
 import { useInterval } from '@/shared/hooks/useInterval';
 
@@ -20,12 +21,6 @@ export const PomodoroController = () => {
   } = usePomodoroStore();
 
   useInterval(tick, isRunning ? 1000 : null);
-
-  const formatTime = (seconds: number): string => {
-    const minutes = Math.floor(seconds / 60);
-    const remainingSeconds = seconds % 60;
-    return `${minutes.toString().padStart(2, '0')}:${remainingSeconds.toString().padStart(2, '0')}`;
-  };
 
   const getProgress = (): number => {
     const totalTime = phase === 'study' ? STUDY_TIME : BREAK_TIME;
